Puck.js: Rework IMU FIFO handling
Use the same accelOn() function, but with optional arguments for FIFO.
This means it will be backwards compatible, but with the option of using
FIFO by passing something like:

Puck.accelOn(208, true, 624);

true is for enabling fifo_mode, and 624 is the number of words (2 bytes)
that the FIFO threshold will be set to. So when the FIFO has 624 words
of data, it will result in a accel event.

The data comes as binary data, and in full sample sets of one gyro sample
and one accelerometer sample. The order of the data is gX,gY,gZ,aX,aY,aX.